created packages and divided classes into them
