+++ src/main/java/store/Task.java
@@ -1,3 +1,5 @@
+package store;
+
 public class Task {
     protected String description;
     protected boolean isDone;
