+++ src/main/java/store/TaskList.java
@@ -1,3 +1,5 @@
+package store;
+
 import java.util.ArrayList;
 
 public class TaskList {
