+++ src/main/java/ui/Ui.java
@@ -1,6 +1,10 @@
+package ui;
+
+import store.Task;
+
 public class Ui {
     public void showWelcomeMessage() {
-        System.out.println("Hello! I'm Dusty\nHow can I help you?\n_______________");
+        System.out.println("Hello! I'm Dusty :> \nHow can I help you?\n_______________");
     }
 
     public void showExitMessage() {
