+++ text-ui-test/EXPECTED.TXT
@@ -1,4 +1,4 @@
-Hello! I'm Dusty
+Hello! I'm ui.Dusty
 How can I help you?
 _______________
 Got it. I've added this task:
